requirements.txt
git+https://github.com/Rapptz/discord.py
requests
pytaf
python-dotenv
markdownify
python-version
emoji==1.6.3